Refactor form JSX files to extract shared utilities and components
- Add formUtils.js with shared FORM_TIMEZONE, formatDateTime24, formatDateOnly,
  formatDateParts, formatTime, and titleCase (all using Pacific time)
- Add formComponents.jsx with shared Header, Row, SectionHeader, and Field components
- Move shared table/narrative CSS into pdf-forms.css under .form-container .page
  using CSS nesting, replacing per-form duplicated rule blocks
- Refactor Form647F.jsx and Form849B.jsx to import from shared modules; inline
  style blocks now contain only page-height/footer overrides
- Refactor FormCoR.jsx to import shared Header, Field, formatDateParts, and
  formatTime; date formatting now uses Pacific time consistently
- Update FormPreview.jsx import to match renamed FormCoR.jsx
