+++ client/src/forms/FormPreview.jsx
@@ -4,14 +4,14 @@ import { useQuery } from '@tanstack/react-query';
 import { Alert, Center, Loader } from '@mantine/core';
 
 import FormContainer from '../../../server/lib/forms/FormContainer.jsx';
-import CertificateOfReleaseForm from 'care-connect-server/lib/forms/CertificateOfReleaseForm.jsx';
+import FormCoR from 'care-connect-server/lib/forms/FormCoR.jsx';
 import Form849B from 'care-connect-server/lib/forms/Form849B.jsx';
 import Form647F from 'care-connect-server/lib/forms/Form647F.jsx';
 
 // Registry mapping form IDs (from the URL) to their React components.
 // Add new form entries here as they are created.
 const FORM_REGISTRY = {
-  cert: CertificateOfReleaseForm,
+  cert: FormCoR,
   '849b': Form849B,
   '647f': Form647F,
 };
