+++ server/lib/forms/Form849B.jsx
@@ -1,55 +1,7 @@
 import React from 'react';
 import { z } from 'zod';
-
-const FORM_TIMEZONE = 'America/Los_Angeles';
-
-function formatDateTime24 (dateStr) {
-  if (!dateStr) return '';
-  const d = new Date(dateStr);
-  if (isNaN(d.getTime())) return '';
-  const date = d.toLocaleString('en-US', { timeZone: FORM_TIMEZONE, month: '2-digit', day: '2-digit', year: 'numeric' });
-  const time = d.toLocaleString('en-US', { timeZone: FORM_TIMEZONE, hour: '2-digit', minute: '2-digit', hour12: false }).replace('24:', '00:');
-  return `${date} ${time}`;
-}
-
-function formatDate (dateStr) {
-  if (!dateStr) return '';
-  const d = new Date(dateStr);
-  if (isNaN(d.getTime())) return '';
-  // Use UTC to avoid timezone shifting a date-only value
-  return d.toLocaleDateString('en-US', { timeZone: 'UTC', month: '2-digit', day: '2-digit', year: 'numeric' });
-}
-
-function titleCase (str) {
-  if (!str) return '';
-  return str.charAt(0).toUpperCase() + str.slice(1).toLowerCase();
-}
-
-function Row ({ label, value }) {
-  return (
-    <tr>
-      <td className='field-label'>{label}</td>
-      <td className='field-value'>{value || ''}</td>
-    </tr>
-  );
-}
-
-function SectionHeader ({ title }) {
-  return (
-    <tr className='section-header-row'>
-      <td colSpan={2} className='section-header'>{title}</td>
-    </tr>
-  );
-}
-
-function Header () {
-  return (
-    <header>
-      Care<span style={{ color: '#888' }}>Connect</span>{' '}
-      <span style={{ fontWeight: 'bold', color: '#bbb' }}>RESET</span>
-    </header>
-  );
-}
+import { FORM_TIMEZONE, formatDateTime24, formatDateOnly, titleCase } from './formUtils.js';
+import { Header, Row, SectionHeader } from './formComponents.jsx';
 
 function buildNarrative ({ arrestedAt, officerName, subjectFullName, arrivedAtReset, transferredAt, releaseReason }) {
   const t1 = formatDateTime24(arrestedAt) || '[date/time]';
@@ -62,20 +14,19 @@ function buildNarrative ({ arrestedAt, officerName, subjectFullName, arrivedAtRe
   return `At ${t1}, SFPD Officer ${officer} arrested ${person} because they were found to be under the influence of a controlled substance or alcohol in a public location. ${person} was brought to RESET at ${t2} and transferred to Sheriff's Office custody at ${t3}. They were subsequently released from their detention due to: ${reason}.`;
 }
 
-const tableCSS = `
-  /*
-   * pdf-forms.css fixes .page to exactly one page height and pins the footer
-   * with position:absolute. Override both here so this form can grow beyond
-   * one page without clipping content, and so the footer flows inline after
-   * the last row rather than overlapping it.
-   *
-   * !important on height: the pdf-forms.css rule uses CSS nesting which may
-   * not be processed before this style block, so we force the override.
-   * The padding is re-declared here so it isn't lost if the nested @media
-   * screen rule in pdf-forms.css doesn't win the cascade.
-   *
-   * Specificity: .form-container .page.form-849b [0,3,0] > .form-container .page [0,2,0]
-   */
+/*
+ * Override pdf-forms.css page constraints so this form can grow beyond
+ * one page without clipping content, and so the footer flows inline after
+ * the last row rather than overlapping it.
+ *
+ * !important on height: the pdf-forms.css rule uses CSS nesting which may
+ * not be processed before this style block, so we force the override.
+ * The padding is re-declared here so it isn't lost if the nested @media
+ * screen rule in pdf-forms.css doesn't win the cascade.
+ *
+ * Specificity: .form-container .page.form-849b [0,3,0] > .form-container .page [0,2,0]
+ */
+const pageCSS = `
   .form-container .page.form-849b {
     height: auto !important;
     min-height: calc(11in - var(--page-margin-top) - var(--page-margin-bottom));
@@ -93,49 +44,6 @@ const tableCSS = `
       left: auto;
       right: auto;
     }
-  }
-
-  .form-849b .form-table {
-    width: 100%;
-    border-collapse: collapse;
-    margin-bottom: 1em;
-    font-size: 9pt;
-    line-height: 1.4;
-  }
-  .form-849b .form-table td {
-    padding: 2.5pt 5pt;
-    border: 0.5pt solid #bbb;
-    vertical-align: top;
-  }
-  .form-849b .form-table .field-label {
-    font-weight: bold;
-    width: 185pt;
-    color: #333;
-    background: #f5f5f5;
-  }
-  .form-849b .form-table .section-header {
-    font-weight: bold;
-    font-size: 7.5pt;
-    background: #ddd;
-    padding: 2pt 5pt;
-    letter-spacing: 0.4pt;
-    text-transform: uppercase;
-  }
-  .form-849b .narrative-section {
-    margin-top: 1em;
-    page-break-inside: avoid;
-  }
-  .form-849b .narrative-label {
-    font-weight: bold;
-    font-size: 9pt;
-    margin-bottom: 3pt;
-  }
-  .form-849b .narrative-text {
-    border: 0.5pt solid #bbb;
-    padding: 6pt;
-    min-height: 65pt;
-    font-size: 9pt;
-    line-height: 1.6;
   }
 `;
 
@@ -261,7 +169,7 @@ export default function Form849B ({ data = {} }) {
 
   return (
     <>
-      <style dangerouslySetInnerHTML={{ __html: tableCSS }} />
+      <style dangerouslySetInnerHTML={{ __html: pageCSS }} />
       <div className='page form-849b'>
         <Header />
 
@@ -297,7 +205,7 @@ export default function Form849B ({ data = {} }) {
             <Row label='Name' value={subjectName} />
             <Row label='Race' value={titleCase(subjectRace)} />
             <Row label='Sex' value={titleCase(subjectSex)} />
-            <Row label='DOB' value={formatDate(subjectDOB)} />
+            <Row label='DOB' value={formatDateOnly(subjectDOB)} />
             <Row label='Alias' value='' />
             <Row label='Height, weight, hair, eyes' value='' />
             <Row label='Address' value={subjectAddress} />
